requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
dataclasses-json==0.5.14
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
langchain==0.0.292
langsmith==0.0.92
marshmallow==3.26.1
multidict==6.7.0
mypy-extensions==1.1.0
numexpr==2.14.1
numpy==1.26.4
openai==0.28.0
packaging==25.0
propcache==0.4.1
pydantic==2.12.4
pydantic-core==2.41.5
python-dotenv==1.0.0
pyyaml==6.0.3
requests==2.32.5
slack-bolt==1.18.0
slack-sdk==3.37.0
sqlalchemy==2.0.44
tenacity==8.5.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.5.0
yarl==1.22.0
